Use a manifest module instead of cartridge headers for launcher menu
Add a script to compile PNGs into byte sequences displayable on the
T3 display.
Load and use the result in the launcher, instead of examining files.

This enables game modules to be compiled to bytecode -- an operation
that loses comments.

Also, bypass the provided Python wrapper for the NeoPixel driver.

Docs pending.